pyproject.toml
[tool.poetry]
name = "text-fetch-etl"
version = "0.1.0"
description = ""
authors = ["cancamilo <camiloramirezf@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = {extras = ["srv"], version = "^4.7.1"}
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
selenium = "^4.18.1"
webdriver-manager = "^4.0.1"
aws-lambda-powertools = "^2.35.1"
beautifulsoup4 = "^4.12.3"
jmespath = "^1.0.1"
pika = "^1.3.2"
telethon = "^1.35.0"
newsapi-python = "^0.2.7"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
